+++ R/subset.R
@@ -61,7 +61,9 @@
   ordered <- is_ordered(x)
   if (!missing(i)) {
     # ordered <- row_validate(i)
-    exceed_rows(result, max(i))
+    if (!is_empty(i)) {
+      exceed_rows(result, max(i))
+    }
     result <- purrr::map(result, `[`, i)
     nr <- length(result[[1]])
     if (!is_min_gap_one(i)) int <- NULL