+++ tests/testthat/test-fill-na.R
@@ -12,9 +12,9 @@ test_that("a tbl_df/data.frame", {
 
 test_that("unknown interval", {
   tsbl <- as_tsibble(dat_x[1, ], index = date)
-  expect_error(fill_na(tsbl), "data of unknown interval.")
-  expect_error(count_gaps(tsbl), "data of unknown interval.")
-  expect_error(has_gaps(tsbl), "data of unknown interval.")
+  expect_identical(fill_na(tsbl), tsbl)
+  expect_equal(count_gaps(tsbl)$n, 0)
+  expect_equal(has_gaps(tsbl), FALSE)
 })
 
 test_that("an irregular tbl_ts", {